Fix printing function in test framework (was segfaulting)
Segfault was occuring in libc heap allocation while rendering float.
I don't know why that segfaults.
Also the test lib tries to avoid allocations, but the author
perhaps didn't realise that libc allocates to render floats?
Converting the elapsed time to an integer gets rid of the allocation,
and also the segfault.